refactor: remover redundant code, regroup folders, move native constants to helpers
