--- android/src/main/java/com/reactnativeescposprinter/ThePrinterState.java
@@ -1,9 +1,0 @@
-package com.reactnativeescposprinter;
-
-public enum ThePrinterState {
-    PRINTER_IDLE,
-    PRINTER_CONNECTING,
-    PRINTER_DISCONNECTING,
-    PRINTER_PRINTING,
-    PRINTER_REMOVING
-}
