+++ ios/EposStringHelper.h
@@ -4,6 +4,10 @@
 @interface EposStringHelper : NSObject
 
 + (NSDictionary *)convertStatusInfoToDictionary:(Epos2PrinterStatusInfo *)status;
++ (NSDictionary *)getDiscoveryConstants;
++ (NSDictionary *)getPrinterConstants;
++ (NSString *)convertDictionatyToJsonString:(NSDictionary *)dict;
 + (int)getPrinterSeries:(NSString*)name;
++ (int)getInitErrorResultCode;
 
 @end
