--- ios/ErrorManager.h
@@ -1,7 +1,0 @@
-#import <Foundation/Foundation.h>
-
-@interface ErrorManager : NSObject
-
-+ (NSString *)convertDictionatyToJsonString:(NSDictionary *)dict;
-
-@end
