+++ ios/EscPosPrinterDiscovery.m
@@ -1,4 +1,5 @@
 #import "EscPosPrinterDiscovery.h"
+#import "EposStringHelper.h"
 
 #import "ePOS2.h"
 
@@ -12,39 +13,11 @@ RCT_EXPORT_MODULE()
 
 - (NSDictionary *)constantsToExport
 {
- return @{
-           // filter options
-          @"PORTTYPE_ALL": @(EPOS2_PORTTYPE_ALL),
-          @"PORTTYPE_TCP": @(EPOS2_PORTTYPE_TCP),
-          @"PORTTYPE_BLUETOOTH": @(EPOS2_PORTTYPE_BLUETOOTH),
-          @"PORTTYPE_USB": @(EPOS2_PORTTYPE_USB),
-          @"PORTTYPE_BLUETOOTH_LE": @(EPOS2_PORTTYPE_BLUETOOTH_LE),
-          @"MODEL_ALL": @(EPOS2_MODEL_ALL),
-          @"TYPE_ALL": @(EPOS2_TYPE_ALL),
-          @"TYPE_PRINTER": @(EPOS2_TYPE_PRINTER),
-          @"TYPE_HYBRID_PRINTER": @(EPOS2_TYPE_HYBRID_PRINTER),
-          @"TYPE_DISPLAY": @(EPOS2_TYPE_DISPLAY),
-          @"TYPE_KEYBOARD": @(EPOS2_TYPE_KEYBOARD),
-          @"TYPE_SCANNER": @(EPOS2_TYPE_SCANNER),
-          @"TYPE_SERIAL": @(EPOS2_TYPE_SERIAL),
-          @"TYPE_POS_KEYBOARD": @(EPOS2_TYPE_POS_KEYBOARD),
-          @"TYPE_MSR": @(EPOS2_TYPE_MSR),
-          @"TYPE_GFE": @(EPOS2_TYPE_GFE),
-          @"TYPE_OTHER_PERIPHERAL": @(EPOS2_TYPE_OTHER_PERIPHERAL),
-          // return value
-          @"ERR_PARAM": @(EPOS2_ERR_PARAM),
-          @"ERR_ILLEGAL": @(EPOS2_ERR_ILLEGAL),
-          @"ERR_MEMORY": @(EPOS2_ERR_MEMORY),
-          @"ERR_FAILURE": @(EPOS2_ERR_FAILURE),
-          @"ERR_PROCESSING": @(EPOS2_ERR_PROCESSING),
-              // pair bluetooth device error
-          @"BT_ERR_PARAM": @(EPOS2_BT_ERR_PARAM),
-          @"BT_ERR_UNSUPPORTED": @(EPOS2_BT_ERR_UNSUPPORTED),
-          @"BT_ERR_CANCEL": @(EPOS2_BT_ERR_CANCEL),
-          @"BT_ERR_ILLEGAL_DEVICE": @(EPOS2_BT_ERR_ILLEGAL_DEVICE),
+ return [EposStringHelper getDiscoveryConstants];
+}
 
-
- };
++ (BOOL) requiresMainQueueSetup {
+  return YES;
 }
 
 - (NSArray<NSString *> *)supportedEvents {
