+++ src/printer/utils/parsers/index.ts
@@ -1,0 +1,2 @@
+export { parsePrinterSettings } from './parsePrinterSettings';
+export { parsePrinterStatus } from './parsePrinterStatus';
